--- types/tbd/openeox.go
@@ -1,9 +1,0 @@
-// SPDX-FileCopyrightText: Copyright 2025 Carabiner Systems, Inc
-// SPDX-License-Identifier: Apache-2.0
-
-package tbd
-
-const (
-	Schema     = "https://docs.oasis-open.org/openeox/tbd/schema/shell.json"
-	CoreSchema = "https://docs.oasis-open.org/openeox/tbd/schema/core.json"
-)
