--- types/tbd/testdata/sample.eox.json
@@ -1,29 +1,0 @@
-{
-  "$schema": "https://docs.oasis-open.org/openeox/tbd/schema/shell.json",
-  "statements": [
-    {
-      "core": {
-        "$schema": "https://docs.oasis-open.org/openeox/tbd/schema/core.json",
-        "end_of_life": "2027-12-31T23:59:59Z",
-        "end_of_security_support": "2027-06-30T23:59:59Z",
-        "end_of_sales": "2026-12-31T23:59:59Z",
-        "last_updated": "2025-04-30T10:00:00Z"
-      },
-      "productName": "Enterprise Server",
-      "productVersion": "5.2",
-      "vendorName": "Example Technologies"
-    },
-    {
-      "core": {
-        "$schema": "https://docs.oasis-open.org/openeox/tbd/schema/core.json",
-        "end_of_life": "2026-09-30T23:59:59Z",
-        "end_of_security_support": "2026-03-31T23:59:59Z",
-        "end_of_sales": "2025-09-30T23:59:59Z",
-        "last_updated": "2025-04-30T10:00:00Z"
-      },
-      "productName": "Secure Gateway",
-      "productVersion": "3.7",
-      "vendorName": "Example Technologies"
-    }
-  ]
-}
