+++ types/v1/openeox.go
@@ -1,0 +1,13 @@
+// SPDX-FileCopyrightText: Copyright 2025 Carabiner Systems, Inc
+// SPDX-License-Identifier: Apache-2.0
+
+package v1
+
+const (
+	Schema     = "https://docs.oasis-open.org/openeox/tbd/schema/shell.json"
+	CoreSchema = "https://docs.oasis-open.org/openeox/v1.0/schema/core.json"
+
+	ProductSoftwareSchema             = "https://docs.oasis-open.org/openeox/tbd/schema/product_software.json"
+	ProductHardwareSchema             = "https://docs.oasis-open.org/openeox/tbd/schema/product_hardware.json"
+	ProductHardwareWithSoftwareSchema = "https://docs.oasis-open.org/openeox/tbd/schema/product_hardware_with_software.json"
+)
