+++ types/v1/testdata/sample.eox.json
@@ -1,0 +1,62 @@
+{
+  "$schema": "https://docs.oasis-open.org/openeox/tbd/schema/shell.json",
+  "statements": [
+    {
+      "core": {
+        "$schema": "https://docs.oasis-open.org/openeox/v1.0/schema/core.json",
+        "endOfLife": {
+          "timestamp": "2027-12-31T23:59:59Z"
+        },
+        "endOfSecuritySupport": {
+          "timestamp": "2027-06-30T23:59:59Z"
+        },
+        "endOfSales": {
+          "timestamp": "2026-12-31T23:59:59Z"
+        },
+        "lastUpdated": "2025-04-30T10:00:00Z"
+      },
+      "product": {
+        "$schema": "https://docs.oasis-open.org/openeox/tbd/schema/product_software.json",
+        "software": {
+          "productName": "Enterprise Server",
+          "productVersion": "5.2",
+          "vendorName": "Example Technologies"
+        }
+      },
+      "productIdentificationHelper": {
+        "cpe": "cpe:2.3:a:example_technologies:enterprise_server:5.2:*:*:*:*:*:*:*",
+        "purls": [
+          "pkg:generic/example-technologies/enterprise-server@5.2"
+        ]
+      }
+    },
+    {
+      "core": {
+        "$schema": "https://docs.oasis-open.org/openeox/v1.0/schema/core.json",
+        "endOfLife": {
+          "timestamp": "2026-09-30T23:59:59Z"
+        },
+        "endOfSecuritySupport": {
+          "timestamp": "2026-03-31T23:59:59Z"
+        },
+        "endOfSales": {
+          "timestamp": "2025-09-30T23:59:59Z"
+        },
+        "lastUpdated": "2025-04-30T10:00:00Z"
+      },
+      "product": {
+        "$schema": "https://docs.oasis-open.org/openeox/tbd/schema/product_software.json",
+        "software": {
+          "productName": "Secure Gateway",
+          "productVersion": "3.7",
+          "vendorName": "Example Technologies"
+        }
+      },
+      "productIdentificationHelper": {
+        "purls": [
+          "pkg:generic/example-technologies/secure-gateway@3.7"
+        ]
+      }
+    }
+  ]
+}
